实现登录功能的数据库操作

+++ internal/domain/user.go
@@ -1,6 +1,6 @@
 package domain
 
 type User struct {
-	Uid    uint
+	Uid    uint64
 	OpenId string
 }
